Producto y ProductoDto tienen nuevo atributo String imageUrl. Cambios varios.

+++ src/main/java/co/utp/misiontic/g12e1/proyectomodisteria/model/entity/Pedido.java
@@ -26,7 +26,7 @@ public class Pedido {
     private Long idPedido;
     
     @Column(name="Precio", nullable = false)
-    private Integer precio;
+    private Double precio;
 
     
     //-------RELACIONES
